requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
python-docx>=1.1.0
Pillow>=10.0.0
